move claim resolve to distribution result
