+++ x/dao/keeper/challenge_test.go
@@ -41,7 +41,7 @@ func TestGetChallengeRange(t *testing.T) {
 	t.Parallel()
 	keeper, ctx := keepertest.DaoKeeper(t)
 	createNChallenge(keeper, ctx, 10)
-	challenges, err := keeper.GetChallengeRange(ctx, 0)
+	challenges, err := keeper.GetChallengeRange(ctx, 0, 9)
 	assert.NoError(t, err)
 	assert.Equal(t, len(challenges), 10)
 }
